+++ hashview.py
@@ -42,6 +42,14 @@ with app.app_context():
         print('\nPlease make sure that your dependencies are up to date (including installing requests).')
         exit(1)
 
+    try:
+        import flask_bcrypt
+        if ('1.0.1' < flask_bcrypt.__version__):
+            raise Exception('old version')
+    except:
+        print('\nPlease make sure that your dependencies are up to date (including replacing Flask-Bcrypt with Bcrypt-Flask).')
+        exit(1)
+
     # If no admins exist prompt user to generate new admin account
     if users == 0:
         print('\nInitial setup detected. Hashview will now prompt you to setup an Administrative account.\n')