+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Notable changes will be documented here
 - Hashview Agent is now packaged with hashview (server) under install directory. Can be downloaded from agents menu as admin
 - Changed from itsdangerous to authlib for password reset token generation (please make sure to update your environments to include authlib).
 - Changed from the python-pushover package, to a call directly to the Pushover API.
+- Changed from Flask-Bcrypt to Bcrypt-Flask.
 ### Removed
 - Removed hashview agent from local processing. If you want to run hashview AND crack hashes on the same box run the hashview-agent in a seperate screen/tmux session